CLI/requirements.txt
numpy
matplotlib
pydicom
scipy
pylinac
pyyaml
